reorganized discourse code into multiple files
